chore : #4 패키지 ê이름 수정, 간단한 URI 수정
