--- lpick-backend/src/main/java/com/notfound/lpickbackend/UserInfo/Query/Repository/UserInfoQueryRepository.java
@@ -1,9 +1,0 @@
-package com.notfound.lpickbackend.UserInfo.Query.Repository;
-
-import com.notfound.lpickbackend.AUTO_ENTITIES.UserInfo;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserInfoQueryRepository extends JpaRepository<UserInfo, String> {
-}
