--- lpick-backend/src/main/java/com/notfound/lpickbackend/Wiki/Command/Application/DTO/Request/PageRevisionRequest.java
@@ -1,15 +1,0 @@
-package com.notfound.lpickbackend.Wiki.Command.Application.DTO.Request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-public class PageRevisionRequest {
-    private String wikiId;
-    private String content;
-}
