--- lpick-backend/src/main/java/com/notfound/lpickbackend/Wiki/Command/Repository/PageRevisionCommandRepository.java
@@ -1,9 +1,0 @@
-package com.notfound.lpickbackend.Wiki.Command.Repository;
-
-import com.notfound.lpickbackend.AUTO_ENTITIES.PageRevision;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PageRevisionCommandRepository extends JpaRepository<PageRevision, String> {
-}
