--- lpick-backend/src/main/java/com/notfound/lpickbackend/Wiki/Query/Repository/WikiPageQueryRepository.java
@@ -1,8 +1,0 @@
-package com.notfound.lpickbackend.Wiki.Query.Repository;
-
-import com.notfound.lpickbackend.AUTO_ENTITIES.WikiPage;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface WikiPageQueryRepository extends JpaRepository<WikiPage, String> {
-
-}
