--- lpick-backend/src/main/java/com/notfound/lpickbackend/_CQRS/ReadME.md
@@ -1,3 +1,0 @@
-### 본 패키지 목적
-패키지 확장 용이 목적으로, CQRS 패턴 구조로 만들어진 비어있는 패키지를 구현해두었습니다.
-만약 추가적인 도메인 확장이나 세분화가 일어날 경우 본 패키지 내역을 복사-붙여넣기 하면 됩니다.
