+++ lpick-backend/src/main/java/com/notfound/lpickbackend/userinfo/config/SecurityConfig.java
@@ -1,0 +1,25 @@
+package com.notfound.lpickbackend.userinfo.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
+
+@Configuration
+@EnableWebSecurity
+public class SecurityConfig {
+
+    @Bean
+    SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+        http
+                // 모든 요청 허용
+                .authorizeHttpRequests(auth -> auth.requestMatchers(new AntPathRequestMatcher("/**")).permitAll())
+                .csrf().disable();
+
+        // SecurityFilterChain을 빌드 후 반환
+        return http.build();
+    }
+
+}
