+++ lpick-backend/src/main/java/com/notfound/lpickbackend/userinfo/query/repository/UserInfoQueryRepository.java
@@ -1,0 +1,9 @@
+package com.notfound.lpickbackend.userinfo.query.repository;
+
+import com.notfound.lpickbackend.AUTO_ENTITIES.UserInfo;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserInfoQueryRepository extends JpaRepository<UserInfo, String> {
+}
