+++ lpick-backend/src/main/java/com/notfound/lpickbackend/wiki/command/application/dto/request/PageRevisionRequest.java
@@ -1,0 +1,15 @@
+package com.notfound.lpickbackend.wiki.command.application.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+public class PageRevisionRequest {
+    private String wikiId;
+    private String content;
+}
