--- lpick-backend/src/main/java/com/notfound/lpickbackend/wikipage/command/application/domain/Status.java
@@ -1,7 +1,0 @@
-package com.notfound.lpickbackend.wikipage.command.application.domain;
-
-public enum Status {
-
-    /* 상태 어떻게 작성하기로 했는지 기억 안나서 임시로 넣는 ENUM 타입 */
-    OPEN
-}
