+++ lpick-backend/src/main/java/com/notfound/lpickbackend/wikipage/command/application/service/WikiPageAndRevisionCommandService.java
@@ -2,9 +2,9 @@ package com.notfound.lpickbackend.wikipage.command.application.service;
 
 
 import com.notfound.lpickbackend.AUTO_ENTITIES.UserInfo;
-import com.notfound.lpickbackend.UserInfo.Query.Service.UserInfoQueryService;
-import com.notfound.lpickbackend.Wiki.Command.Application.DTO.Request.PageRevisionRequest;
-import com.notfound.lpickbackend.Wiki.Command.Application.Service.PageRevisionCommandService;
+import com.notfound.lpickbackend.userinfo.query.service.UserInfoQueryService;
+import com.notfound.lpickbackend.wiki.command.application.dto.request.PageRevisionRequest;
+import com.notfound.lpickbackend.wiki.command.application.service.PageRevisionCommandService;
 import com.notfound.lpickbackend.wikipage.command.application.dto.WikiPageCreateRequestDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
