+++ lpick-backend/src/main/java/com/notfound/lpickbackend/wikipage/command/application/service/WikiPageCommandService.java
@@ -1,9 +1,9 @@
 package com.notfound.lpickbackend.wikipage.command.application.service;
 
-import com.notfound.lpickbackend.AUTO_ENTITIES.WikiPage;
+import com.notfound.lpickbackend.wikipage.command.application.domain.WikiPage;
 import com.notfound.lpickbackend.common.exception.CustomException;
 import com.notfound.lpickbackend.common.exception.ErrorCode;
-import com.notfound.lpickbackend.wikipage.command.application.domain.Status;
+import com.notfound.lpickbackend.wikipage.command.application.domain.WikiStatus;
 import com.notfound.lpickbackend.wikipage.command.repository.WikiPageCommandRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -26,7 +26,7 @@ public class WikiPageCommandService {
             WikiPage newWikiPage = WikiPage.builder()
                     .title(title)
                     .currentRevision("r1") // 초기 생성시 r1
-                    .status(Status.OPEN)
+                    .wikiStatus(WikiStatus.OPEN)
                     .build();
 
             wikiPageCommandRepository.save(newWikiPage);
