+++ lpick-backend/src/main/java/com/notfound/lpickbackend/wikipage/command/repository/WikiPageCommandRepository.java
@@ -1,6 +1,6 @@
 package com.notfound.lpickbackend.wikipage.command.repository;
 
-import com.notfound.lpickbackend.AUTO_ENTITIES.WikiPage;
+import com.notfound.lpickbackend.wikipage.command.application.domain.WikiPage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
